BIG update! It works!
This version is capable of decrypting simple ciphers, such as the Caesar cipher.
You need to train a model first! Don't worry, it's easy. 
Load some cleartext (the more the better) into the program, then use one of the options to create a model.
Then, restart the program, and load in your chosen ciphertext file... and get decrypting!
In my own tests, it was able to decrypt a simple Caesar cipher with 92% accuracy, without actually 'knowing' what cipher was being used.